+++ db/migrate/20260601085213_add_sms_provider_to_users.rb
@@ -1,0 +1,10 @@
+class AddSmsProviderToUsers < ActiveRecord::Migration[8.1]
+  def change
+    add_column :users, :sms_provider, :string, default: "aws", null: false
+
+    execute <<~SQL # rubocop:disable Rails/ReversibleMigration
+      UPDATE users SET sms_provider = 'twilio'
+      WHERE created_at < '2026-04-01'
+    SQL
+  end
+end